requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9; platform_system != 'Windows'
python-dateutil>=2.9
django-environ>=0.11
kafka-python>=2.0
elasticsearch>=8.15
uvicorn>=0.30
orjson>=3.10
python-json-logger>=2.0
rapidfuzz>=3.9
gunicorn>=21.2